pyproject.toml
# ==============================================================================
# This file is part of the WISDAM distribution
# https://github.com/WISDAMapp/WISDAM
# Copyright (C) 2024 Martin Wieser.
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see http://www.gnu.org/licenses/.
# ==============================================================================


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "WISDAM"
version = "1.0.6"
authors = [
  { name="Martin Wieser"},
]
description = "A software to work with imagery from imagery pilotet aircrafts, drones or orthophotos"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: Microsoft :: Windows :: Windows 10",
	"Operating System :: Microsoft :: Windows :: Windows 11",
    "Development Status :: 3 - Alpha"
]

dependencies = [
    'numpy>=1.26.0',
	'pyproj==3.6.1',
	'shapely>=2.0.1',
	'fiona>=1.9.5',
	'pandas>=2.1.1',
	'natsort>=8.4.0',
	'openpyxl>=3.1.3',
	'geopandas>=0.14.4',
	'Pillow>=10.0.1',
	'toml>=0.10.2',
	'rawpy>=0.18.1',
	'PySide6>=6.8.0',
	'PySide6_Essentials>=6.8.0',
	'PySide6_Addons>=6.8.0',
	'PyExifTool>=0.5.6']

[project.optional-dependencies]
test = [
    'pytest>=8.0.0'
]

[project.urls]
    "Homepage" = "https://github.com/WISDAMapp/WISDAM"
    "Bug Tracker" = "https://github.com/WISDAMapp/WISDAM/issues"


[project.scripts]
wisdam = "WISDAM.main:run"